Context aware header
You can now choose between 2 headers.
The Logo header will just display the selected logo.
The context Aware header will show a logo that corrsponds to the
contexr. e.g. when in a job it will show the job status or the
weather, dito for folders.
On manage Jenkins pages the plugin tries to display a fitting symbol
from ionicons.
TODO: make the symbol mapping configurable via a file

Solve #4: allow html in the title