add a page decorator
the page decorator allows us to load the css and javascript as part of
the <head> section of the html document. Before it was only loaded as
part of the html body which seems to lead to a noticable delay before
styles are applied.A
Additionally the css is preloaded also an image if used is preloaded.