fix hanging request
when using a SVGLogo and providing a non existent path or a non existent
url leads to a hanging ui as the http request seems to hang forever.
The cause are the new 404 pages, that include the header itself and a
missing timeout on the connection.
Properly handle the return code when loading from a url
Increase the caching period to 1 hour