Add `npm` installation
In trying to add this plugin to [bom](https://github.com/jenkinsci/bom), I was unable to run:

`mvn clean verify`

locally to do testing since I did not have `npm` installed:

```
[INFO] --- frontend:1.15.0:yarn (yarn mvnbuild) @ customizable-header ---
[INFO] Running 'yarn run mvnbuild' in /Users/dpope/github-forks/customizable-header-plugin
[INFO] yarn run v1.22.19
[INFO] $ npm run build
[INFO] /bin/sh: npm: command not found
[INFO] error Command failed with exit code 127.
[INFO] error Command failed with exit code 127.info Visit https://yarnpkg.com/en/docs/cli/run for documentation about this command.
```

There may be a better way to solve this issue, but this was the simplest way for me.

Once this issue is solved (in whatever way you prefer), I can add this plugin to bom.