add classic search as suffix
Bring back a search box in the header. The difference to command palette
is that when one presses enter while the input has focus it will get you
to the old search output showing the list of results.
Similar to the command palette it displays an icon,
Compared to the old (YUI based) search the results are links that can be
clicked directly.