Fix plugin loading when datadog is installed
with datadog plugin installed the call to save in readResolve triggers
the saveable listener of the datadog plugin which then results in
loading the current user. This tries to read the header configuration
which is not yet properly initialized. The effect is that the header
config is not there later breaking the UI completely.