improve places that use text-color-secondary
The header overwrites the text-color-secondary variable.
Dependeing on theme and header color the placeholder and the separator
in the app-nav menu become unreadable.
For placeholder use a color based on text-color.
For the app-nav menu attach it to the body instead of the header. This avoids
that the secondary color from the header is used.